code working together, added new dependency to solve dependency issue. Now 500 error, otherwise working?
