+++ .idea/compiler.xml
@@ -10,6 +10,9 @@
       <module name="demo.test" target="17" />
       <module name="IndigenoUS-Seed-Exchange-Network" target="1.7" />
       <module name="IndigenoUS-Seed-Network-Exchange" target="1.7" />
+      <module name="IndigenoUS_Seed_Exchange_Network" target="17" />
+      <module name="IndigenoUS_Seed_Exchange_Network.main" target="17" />
+      <module name="IndigenoUS_Seed_Exchange_Network.test" target="17" />
       <module name="IndigenoUS_Seed_Network_Exchange" target="17" />
       <module name="IndigenoUS_Seed_Network_Exchange.main" target="17" />
       <module name="IndigenoUS_Seed_Network_Exchange.test" target="17" />
