--- .idea/libraries/Gradle__com_google_protobuf_protobuf_java_3_11_4.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: com.google.protobuf:protobuf-java:3.11.4" type="java-imported" external-system-id="GRADLE">
-    <properties groupId="com.google.protobuf" artifactId="protobuf-java" version="3.11.4" />
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.google.protobuf/protobuf-java/3.11.4/7ec0925cc3aef0335bbc7d57edfd42b0f86f8267/protobuf-java-3.11.4.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.google.protobuf/protobuf-java/3.11.4/1e3a86dde8ecf9410e23f22a67ad062316fe94a/protobuf-java-3.11.4-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
