--- .idea/libraries/Gradle__mysql_mysql_connector_java_8_0_28.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: mysql:mysql-connector-java:8.0.28" type="java-imported" external-system-id="GRADLE">
-    <properties groupId="mysql" artifactId="mysql-connector-java" version="8.0.28" />
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/mysql/mysql-connector-java/8.0.28/33678b1729d4f832b9e4bcb2d5bbd67940920a7a/mysql-connector-java-8.0.28.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/mysql/mysql-connector-java/8.0.28/de25e9f0fc9ae5e9273e59eecde038096197809/mysql-connector-java-8.0.28-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
