--- .idea/libraries/Gradle__org_springframework_boot_spring_boot_starter_mail_3_2_2.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: org.springframework.boot:spring-boot-starter-mail:3.2.2" type="java-imported" external-system-id="GRADLE">
-    <properties groupId="org.springframework.boot" artifactId="spring-boot-starter-mail" version="3.2.2" />
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.springframework.boot/spring-boot-starter-mail/3.2.2/22caa13d58241fa44466edc2a2587992a203791e/spring-boot-starter-mail-3.2.2.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.springframework.boot/spring-boot-starter-mail/3.2.2/da38855633fe58ddb80777f604b7190a1d200b44/spring-boot-starter-mail-3.2.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
