--- .idea/libraries/Gradle__org_springframework_security_spring_security_crypto_5_5_8.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: org.springframework.security:spring-security-crypto:5.5.8" type="java-imported" external-system-id="GRADLE">
-    <properties groupId="org.springframework.security" artifactId="spring-security-crypto" version="5.5.8" />
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.springframework.security/spring-security-crypto/5.5.8/4d0d3c5f49f94c1b75431ae426f4a3e351b09c91/spring-security-crypto-5.5.8.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.springframework.security/spring-security-crypto/5.5.8/acfb9547e9873903a35adbc5573096b4f8682d4c/spring-security-crypto-5.5.8-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
