--- .idea/libraries/Gradle__org_springframework_spring_context_support_6_1_3.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: org.springframework:spring-context-support:6.1.3" type="java-imported" external-system-id="GRADLE">
-    <properties groupId="org.springframework" artifactId="spring-context-support" version="6.1.3" />
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.springframework/spring-context-support/6.1.3/55ff73910c4851c37f6919a071453ddc40f7dc38/spring-context-support-6.1.3.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.springframework/spring-context-support/6.1.3/8dd69cd7e0e85a67b75f45dcaeef1b7704640917/spring-context-support-6.1.3-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
