+++ .idea/modules.xml
@@ -2,9 +2,9 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/demo.iml" filepath="$PROJECT_DIR$/.idea/modules/demo.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/demo.main.iml" filepath="$PROJECT_DIR$/.idea/modules/demo.main.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/demo.test.iml" filepath="$PROJECT_DIR$/.idea/modules/demo.test.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/IndigenoUS_Seed_Exchange_Network.iml" filepath="$PROJECT_DIR$/.idea/modules/IndigenoUS_Seed_Exchange_Network.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/IndigenoUS_Seed_Exchange_Network.main.iml" filepath="$PROJECT_DIR$/.idea/modules/IndigenoUS_Seed_Exchange_Network.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/IndigenoUS_Seed_Exchange_Network.test.iml" filepath="$PROJECT_DIR$/.idea/modules/IndigenoUS_Seed_Exchange_Network.test.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/group-michael-1.iml" filepath="$PROJECT_DIR$/.idea/group-michael-1.iml" />
     </modules>
   </component>
