+++ .idea/modules/IndigenoUS_Seed_Exchange_Network.iml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="demo" external.linked.project.path="$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network" external.root.project.path="$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network" external.system.id="GRADLE" external.system.module.group="org.launchcode" external.system.module.version="0.0.1-SNAPSHOT" type="JAVA_MODULE" version="4">
+<module external.linked.project.id="IndigenoUS_Seed_Exchange_Network" external.linked.project.path="$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network" external.root.project.path="$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network" external.system.id="GRADLE" external.system.module.group="org.launchcode" external.system.module.version="0.0.1-SNAPSHOT" type="JAVA_MODULE" version="4">
   <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_17" inherit-compiler-output="true">
     <exclude-output />
     <content url="file://$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network">
