+++ .idea/modules/IndigenoUS_Seed_Exchange_Network.test.iml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="demo:test" external.linked.project.path="$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network" external.root.project.path="$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network" external.system.id="GRADLE" external.system.module.group="org.launchcode" external.system.module.type="sourceSet" external.system.module.version="0.0.1-SNAPSHOT" type="JAVA_MODULE" version="4">
+<module external.linked.project.id="IndigenoUS_Seed_Exchange_Network:test" external.linked.project.path="$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network" external.root.project.path="$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network" external.system.id="GRADLE" external.system.module.group="org.launchcode" external.system.module.type="sourceSet" external.system.module.version="0.0.1-SNAPSHOT" type="JAVA_MODULE" version="4">
   <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_17">
     <output-test url="file://$MODULE_DIR$/../../IndigenoUS_Seed_Exchange_Network/build/classes/java/test" />
     <exclude-output />
@@ -8,7 +8,7 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module" module-name="demo.main" />
+    <orderEntry type="module" module-name="IndigenoUS_Seed_Exchange_Network.main" />
     <orderEntry type="library" name="Gradle: org.springframework.boot:spring-boot-starter-data-jpa:3.2.2" level="project" />
     <orderEntry type="library" name="Gradle: org.springframework.boot:spring-boot-starter-mail:3.2.2" level="project" />
     <orderEntry type="library" name="Gradle: org.springframework.security:spring-security-crypto:5.5.8" level="project" />
@@ -17,6 +17,7 @@
     <orderEntry type="library" name="Gradle: org.springframework.boot:spring-boot-starter-web:3.2.2" level="project" />
     <orderEntry type="library" name="Gradle: mysql:mysql-connector-java:8.0.28" level="project" />
     <orderEntry type="library" name="Gradle: com.h2database:h2:2.2.224" level="project" />
+    <orderEntry type="library" name="Gradle: jakarta.mail:jakarta.mail-api:2.0.1" level="project" />
     <orderEntry type="library" name="Gradle: org.springframework.boot:spring-boot-starter-test:3.2.2" level="project" />
     <orderEntry type="library" name="Gradle: org.springframework.boot:spring-boot-starter-aop:3.2.2" level="project" />
     <orderEntry type="library" name="Gradle: org.springframework.boot:spring-boot-starter-jdbc:3.2.2" level="project" />
@@ -113,5 +114,5 @@
     <orderEntry type="library" scope="RUNTIME" name="Gradle: org.glassfish.jaxb:txw2:4.0.4" level="project" />
     <orderEntry type="library" scope="RUNTIME" name="Gradle: com.sun.istack:istack-commons-runtime:4.1.2" level="project" />
   </component>
-  <component name="TestModuleProperties" production-module="demo.main" />
+  <component name="TestModuleProperties" production-module="IndigenoUS_Seed_Exchange_Network.main" />
 </module>
