+++ IndigenoUS_Seed_Exchange_Network/build.gradle.kts
@@ -1,41 +1,37 @@
 plugins {
-    java
-    id("org.springframework.boot") version "3.2.2"
-    id("io.spring.dependency-management") version "1.1.4"
+	java
+	id("org.springframework.boot") version "3.2.2"
+	id("io.spring.dependency-management") version "1.1.4"
 }
 
 group = "org.launchcode"
 version = "0.0.1-SNAPSHOT"
 
 java {
-
-	sourceCompatibility = JavaVersion.VERSION_17
-
-
+	sourceCompatibility = '17'
 }
 
 repositories {
 	mavenCentral()
 }
 
-
 dependencies {
-
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-    implementation("org.springframework.boot:spring-boot-starter-mail")
-    implementation("org.springframework.security:spring-security-crypto:5.5.8")
-    implementation("org.springframework.boot:spring-boot-starter-thymeleaf")
-    implementation("org.springframework.boot:spring-boot-starter-validation")
-    implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("mysql:mysql-connector-java:8.0.28")
-    implementation ("com.h2database:h2")
-    developmentOnly("org.springframework.boot:spring-boot-devtools")
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
+	implementation("org.springframework.boot:spring-boot-starter-data-jpa")
+	implementation("org.springframework.boot:spring-boot-starter-mail")
+	implementation("org.springframework.security:spring-security-crypto:5.5.8")
+	implementation("org.springframework.boot:spring-boot-starter-thymeleaf")
+	implementation("org.springframework.boot:spring-boot-starter-validation")
+	implementation("org.springframework.boot:spring-boot-starter-web")
+	implementation("mysql:mysql-connector-java:8.0.28")
+	implementation ("com.h2database:h2")
+	developmentOnly("org.springframework.boot:spring-boot-devtools")
+	testImplementation("org.springframework.boot:spring-boot-starter-test")
+	implementation ("jakarta.mail:jakarta.mail-api:2.0.1")
 
 }
 
 tasks.withType<Test> {
-    useJUnitPlatform()
+	useJUnitPlatform()
 }
 
 //Unit 2, Class 12 abt 17:00 dependency import tip
