+++ IndigenoUS_Seed_Exchange_Network/src/main/java/org/launchcode/IndigenoUS_Seed_Exchange_Network/data/UserRepository.java
@@ -1,0 +1,13 @@
+package org.launchcode.IndigenoUS_Seed_Exchange_Network.data;
+
+import org.launchcode.IndigenoUS_Seed_Exchange_Network.models.ProfessionalReference;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends CrudRepository<ProfessionalReference.User, Integer> {
+    Optional<ProfessionalReference.User> findByEmail(String email);
+}
+
