+++ IndigenoUS_Seed_Exchange_Network/src/main/java/org/launchcode/IndigenoUS_Seed_Exchange_Network/models/MailConfiguration.java
@@ -1,4 +1,4 @@
-package org.launchcode;
+package org.launchcode.IndigenoUS_Seed_Exchange_Network.models;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
