--- IndigenoUS_Seed_Exchange_Network/src/main/java/org/launchcode/IndigenoUsSeedExchangeNetworkApplication.java
@@ -1,13 +1,0 @@
-package org.launchcode;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class IndigenoUsSeedExchangeNetworkApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(IndigenoUsSeedExchangeNetworkApplication.class, args);
-	}
-
-}
