--- IndigenoUS_Seed_Exchange_Network/src/main/java/org/launchcode/model/ProfessionalReference.java
@@ -1,40 +1,0 @@
-package org.launchcode.model;
-
-public class ProfessionalReference {
-    private String name;
-    private String institution;
-    private String position;
-    private String email;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getInstitution() {
-        return institution;
-    }
-
-    public void setInstitution(String institution) {
-        this.institution = institution;
-    }
-
-    public String getPosition() {
-        return position;
-    }
-
-    public void setPosition(String position) {
-        this.position = position;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
