--- IndigenoUS_Seed_Exchange_Network/src/main/java/org/launchcode/model/User.java
@@ -1,50 +1,0 @@
-package org.launchcode.model;
-
-
-import jakarta.persistence.*;
-
-
-@Entity
-@Table(name = "users", uniqueConstraints = @UniqueConstraint(columnNames = "email"))
-public class User {
-    
-    @Id
-    @GeneratedValue
-
-    private Integer id;
-    private String email;
-    private String password;
-
-    public User(){}
-
-    public User(Integer id, String fullName, String email, String password, String role) {
-        this.id = id;
-        this.email = email;
-        this.password = password;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
