--- IndigenoUS_Seed_Exchange_Network/src/main/java/org/launchcode/model/data/UserRepository.java
@@ -1,13 +1,0 @@
-package org.launchcode.model.data;
-
-import org.launchcode.model.User;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
-    Optional<User> findByEmail(String email);
-}
-
