+++ IndigenoUS_Seed_Exchange_Network/src/main/resources/templates/application-form.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org/">
 <head>
+    <p th:replace="fragments::navbar"></p>
     <meta charset="UTF-8">
     <title>Application Form</title>
     <style>
@@ -76,29 +77,29 @@
 
 
     <div class="form-group">
-        <label for="refName2">Reference 2 Name:</label>
-        <input type="text" id="refName2" name="refName2" th:field="*{reference2.name}" class="form-control" required>
+         <label for="refName2">Reference 2 Name:</label>
+         <input type="text" id="refName2" name="refName2" th:field="*{reference2.name}" class="form-control" required>
 
-        <label for="refInstitution2">Institution:</label>
-        <input type="text" id="refInstitution2" name="refInstitution2" th:field="*{reference2.institution}" class="form-control" required>
+         <label for="refInstitution2">Institution:</label>
+         <input type="text" id="refInstitution2" name="refInstitution2" th:field="*{reference2.institution}" class="form-control" required>
 
-        <label for="refPosition2">Position:</label>
-        <input type="text" id="refPosition2" name="refPosition2" th:field="*{reference2.position}" class="form-control" required>
+         <label for="refPosition2">Position:</label>
+         <input type="text" id="refPosition2" name="refPosition2" th:field="*{reference2.position}" class="form-control" required>
 
-        <label for="refEmail2">Email:</label>
-        <input type="email" id="refEmail2" name="refEmail2" th:field="*{reference2.email}" class="form-control" required>
-    </div>
-
-
-    <div class="form-group">
-        <button type="submit" class="btn btn-success">Submit</button>
-    </div>
-</form>
+         <label for="refEmail2">Email:</label>
+         <input type="email" id="refEmail2" name="refEmail2" th:field="*{reference2.email}" class="form-control" required>
+     </div>
 
 
-<div th:if="${successMessage}" class="alert alert-success" role="alert" style="max-width:600px;">
-    <p th:text="${successMessage}" class="form-control"></p>
-</div>
+     <div class="form-group">
+         <button type="submit" class="btn btn-success">Submit</button>
+     </div>
+ </form>
 
-</body>
-</html>
+
+ <div th:if="${successMessage}" class="alert alert-success" role="alert" style="max-width:600px;">
+     <p th:text="${successMessage}" class="form-control"></p>
+ </div>
+
+ </body>
+ </html>
