+++ IndigenoUS_Seed_Exchange_Network/src/main/resources/templates/fragments.html
@@ -10,6 +10,7 @@
 <div th:fragment="navbar">
     <a href="/">Blog</a>
     <a href="/new-post">New Post</a>
+    <a href="/application">Application</a>
 </div>
 </body>
 </html>
